fixed pagination standard response

+++ src/stores/connectors/ethereum/list_test.go
@@ -57,7 +57,7 @@ func TestListEthAccounts(t *testing.T) {
 
 	t.Run("should fail to list ethAccounts if db fails", func(t *testing.T) {
 		auth.EXPECT().CheckPermission(&types.Operation{Action: types.ActionRead, Resource: types.ResourceEthAccount}).Return(nil)
-		db.EXPECT().SearchAddresses(gomock.Any(), false, 0, 0).Return(nil, expectedErr)
+		db.EXPECT().SearchAddresses(gomock.Any(), false, uint64(0), uint64(0)).Return(nil, expectedErr)
 
 		_, err := connector.List(ctx, 0, 0)
 
@@ -98,7 +98,7 @@ func TestListDeletedEthAccounts(t *testing.T) {
 	t.Run("should fail with same error if authorization fails", func(t *testing.T) {
 		auth.EXPECT().CheckPermission(&types.Operation{Action: types.ActionRead, Resource: types.ResourceEthAccount}).Return(expectedErr)
 
-		_, err := connector.ListDeleted(ctx, 0, 0)
+		_, err := connector.ListDeleted(ctx, uint64(0), uint64(0))
 
 		assert.Error(t, err)
 		assert.Equal(t, err, expectedErr)
@@ -106,9 +106,9 @@ func TestListDeletedEthAccounts(t *testing.T) {
 
 	t.Run("should fail to list deleted ethAccounts if db fails", func(t *testing.T) {
 		auth.EXPECT().CheckPermission(&types.Operation{Action: types.ActionRead, Resource: types.ResourceEthAccount}).Return(nil)
-		db.EXPECT().SearchAddresses(gomock.Any(), true, 0, 0).Return(nil, expectedErr)
+		db.EXPECT().SearchAddresses(gomock.Any(), true, uint64(0), uint64(0)).Return(nil, expectedErr)
 
-		_, err := connector.ListDeleted(ctx, 0, 0)
+		_, err := connector.ListDeleted(ctx, uint64(0), uint64(0))
 
 		assert.Error(t, err)
 		assert.Equal(t, err, expectedErr)
