+++ src/stores/connectors/keys/list_test.go
@@ -56,9 +56,9 @@ func TestListKey(t *testing.T) {
 
 	t.Run("should fail to list keys if db fails", func(t *testing.T) {
 		auth.EXPECT().CheckPermission(&types.Operation{Action: types.ActionRead, Resource: types.ResourceKey}).Return(nil)
-		db.EXPECT().SearchIDs(gomock.Any(), false, 0, 0).Return(nil, expectedErr)
+		db.EXPECT().SearchIDs(gomock.Any(), false, uint64(0), uint64(0)).Return(nil, expectedErr)
 
-		_, err := connector.List(ctx, 0, 0)
+		_, err := connector.List(ctx, uint64(0), uint64(0))
 
 		assert.Error(t, err)
 		assert.Equal(t, err, expectedErr)
@@ -97,7 +97,7 @@ func TestListDeletedKey(t *testing.T) {
 	t.Run("should fail with same error if authorization fails", func(t *testing.T) {
 		auth.EXPECT().CheckPermission(&types.Operation{Action: types.ActionRead, Resource: types.ResourceKey}).Return(expectedErr)
 
-		_, err := connector.ListDeleted(ctx, 0, 0)
+		_, err := connector.ListDeleted(ctx, uint64(0), uint64(0))
 
 		assert.Error(t, err)
 		assert.Equal(t, err, expectedErr)
@@ -105,9 +105,9 @@ func TestListDeletedKey(t *testing.T) {
 
 	t.Run("should fail to list deleted key if db fails", func(t *testing.T) {
 		auth.EXPECT().CheckPermission(&types.Operation{Action: types.ActionRead, Resource: types.ResourceKey}).Return(nil)
-		db.EXPECT().SearchIDs(gomock.Any(), true, 0, 0).Return(nil, expectedErr)
+		db.EXPECT().SearchIDs(gomock.Any(), true, uint64(0), uint64(0)).Return(nil, expectedErr)
 
-		_, err := connector.ListDeleted(ctx, 0, 0)
+		_, err := connector.ListDeleted(ctx, uint64(0), uint64(0))
 
 		assert.Error(t, err)
 		assert.Equal(t, err, expectedErr)
